toolbox-Effect: Shift All That Can Stay Home
Fedora `toolbox` reads `~`, but not `/etc/`, causing serious
side-effects. I may or may not have lost all my accumulated bash history
to such a side-effect.

This commit gets rid of all config that can be placed in `~` as
dotfiles. Refer to ankitpati/dotfiles to know the fate of these files.